integration_2023_wk06

See merge request oai/openairinterface5g!1943

!1855 Cleanup of external dependencies install script
!1876 NR refactor PHY vector sizes for multi UE
!1933 bugfix: try to handle USRP devices correctly to avoid crashes
!1937 CI: clean unused files, drop old cluster images
!1629 E1 procedure
!1738 Replace Nettle with OpenSSL
!1896 Remove CSET0 rom configuration if not in configuration file

+++ openair2/ENB_APP/enb_config.c
@@ -2015,7 +2015,7 @@ int RCconfig_gtpu(void ) {
   int               num_enbs                      = 0;
   char             *enb_interface_name_for_S1U    = NULL;
   char             *enb_ipv4_address_for_S1U      = NULL;
-  uint32_t          enb_port_for_S1U              = 0;
+  uint16_t enb_port_for_S1U = 0;
   char             *address                       = NULL;
   char             *cidr                          = NULL;
   char gtpupath[MAX_OPTNAME_SIZE*2 + 8];
